requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
git-filter-repo==2.47.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psutil==7.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
